app/pipelines/three_d_generation/components/texture_generator.py
# pipeline/components/texture_generator.py

from PIL import Image
from typing import Tuple, Any
from pathlib import Path


class TextureGenerator3D:
    def __init__(self, model_path: str = "hunyuan3d-paint-v2-0"):
        self.model_path = model_path
        self.texture_model = self.load_model()

    def load_model(self) -> Any:
        # TODO: Load texture generation model
        print(f"Loading texture model from {self.model_path}")
        return None

    def generate_multiview_texture(self, ref_image: Image.Image, normals: Any, positions: Any, camera_config: Any) -> Any:
        # TODO: Generate multiview texture using Hy3DSampleMultiView
        print("Generating multiview texture...")
        return None

    def bake_texture(self, multiviews: Any, renderer: Any, camera_config: Any) -> Tuple[Image.Image, Any, Any]:
        # TODO: Use Hy3DBakeFromMultiview
        print("Baking texture...")
        texture = None
        mask = None
        return texture, mask, renderer

    def inpaint_texture(self, texture: Image.Image, mask: Any, renderer: Any) -> Image.Image:
        # TODO: Use vertex inpaint or CV2 inpainting
        print("Inpainting texture...")
        return texture

    def apply_and_export_texture(self, texture: Image.Image, renderer: Any, output_path: Path) -> Path:
        # TODO: Apply texture to mesh and export to .glb
        print(f"Exporting textured mesh to {output_path}")
        output_path.parent.mkdir(parents=True, exist_ok=True)
        return output_path
