requirements.txt
python-dotenv==1.1.1
pydantic==2.11.7
uvicorn==0.35.0
fastapi==0.116.1
pillow==11.3.0